+++ examples/vtzero-stats.cpp
@@ -37,4 +37,6 @@ int main(int argc, char* argv[]) {
         std::cerr << "Error: " << e.what() << '\n';
         return 1;
     }
+
+    return 0;
 }