+++ examples/vtzero-streets.cpp
@@ -75,5 +75,7 @@ int main(int argc, char* argv[]) {
         std::cerr << "Error: " << e.what() << '\n';
         return 1;
     }
+
+    return 0;
 }